another commit

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +25,8 @@ class PostsController extends Controller
      */
     public function index()
     {
-        $posts= Post::all(); 
+        // $posts= Post::OrderBy('title','asc')->paginate(2); 
+        $posts= Post::all();
         return view('posts.index', compact('posts'));
     }
 
@@ -38,13 +50,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title'=>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image' =>'image|nullable|max:1999'
         ]);
+        // handle file upload
+        if($request->hasFile('cover_image')){
+        // get filename with ext
+        $fileNameWithExt = $request->file('cover_image')->GetClientOriginalName();
+        //get just filename
+        $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+        //get just extension
+        $extension = $request->file('cover_image')->GetClientOriginalExtension();
+        //filename to store
+        $fileNameToStore = $filename.'_'.time().'.'.$extension;
+        //upload image
+        $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore = "noimage.jpg";
+        }
         // create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         // redirecting post
         return redirect('/posts')->with('success','post created');
@@ -72,6 +101,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -88,10 +120,26 @@ class PostsController extends Controller
             'title'=>'required',
             'body'=>'required'
         ]);
+          // handle file upload
+          if($request->hasFile('cover_image')){
+            // get filename with ext
+            $fileNameWithExt = $request->file('cover_image')->GetClientOriginalName();
+            //get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //get just extension
+            $extension = $request->file('cover_image')->GetClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+            }
         // create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         // redirecting post
         return redirect('/posts')->with('success','post updated');
@@ -106,6 +154,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post= Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised page');
+        }
+        if($post->cover_image !='noimage.jpg'){
+            // delete image
+            Strorage::delete('public/cover_images'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success','post removed');
